+++ finished/app/components/EditableElement.jsx
@@ -1,31 +1,47 @@
 var React = require('react');
 
-var EditableElement = React.createClass({
+module.exports = React.createClass({
+
+  displayName: "EditableElement",
+
+  propTypes: {
+    value: React.PropTypes.oneOfType([
+      React.PropTypes.string,
+      React.PropTypes.number,
+    ]).isRequired,
+    onEdit: React.PropTypes.func.isRequired
+  },
+
+  getDefaultProps: function() {
+    return {
+      value: 'Enter New Value'
+    }
+  },
+
   getInitialState: function() {
     return { editing: false };
   },
 
   render: function() {
-    return this.state.editing ? this.renderEdit() : this.renderValue()
+    var isEditing = this.state.editing;
+    return isEditing ? this.renderInputField() : this.renderParagraph()
   },
 
-  renderValue: function() {
-    var props = this.props;
-    return <p onClick={this.enterEditState}>{props.item || props.defaultValue}</p>
-  },
-
-  enterEditState: function() {
-    this.setState({ editing: true });
-  },
-
-  renderEdit: function() {
+  renderInputField: function() {
     return (
       <input type="text"
         autoFocus="true"
-        defaultValue={this.props.item}
         onBlur={this.finishEdit}
         onKeyPress={this.checkEnter} />
     );
+  },
+
+  renderParagraph: function() {
+    return <p onClick={this.enterEditState}>{this.props.value}</p>
+  },
+
+  enterEditState: function() {
+    this.setState({editing: true});
   },
 
   checkEnter: function(e) {
@@ -35,11 +51,8 @@ var EditableElement = React.createClass({
   },
 
   finishEdit: function(e) {
-    var keyName = this.props.keyName;
     var newValue = e.target.value;
-    this.props.onEdit(keyName, newValue);
+    this.props.onEdit(newValue);
     this.setState({editing: false});
   }
 });
-
-module.exports = EditableElement;
