+++ finished/app/components/NavigationDashboard.jsx
@@ -1,7 +1,10 @@
 var React = require('react');
 var nav   = require('../utilities/starshipNavigation.js');
 
-var NavigationDashboard = React.createClass({
+module.exports = React.createClass({
+
+  displayName: "NavigationDashboard",
+
   render: function() {
     var ship = this.props.ship
     var destination = ship.destination;
@@ -20,5 +23,3 @@ var NavigationDashboard = React.createClass({
     );
   }
 });
-
-module.exports = NavigationDashboard;
