+++ finished/app/components/ShipInfo.jsx
@@ -1,38 +1,47 @@
 var React            = require('react');
 var EditableElement  = require('./EditableElement.jsx');
 
-var ShipInfo = React.createClass({
+module.exports = React.createClass({
+
+  displayName: "ShipInfo",
+
+  shouldComponentUpdate: function(nextProps, nextState) {
+    return this.props.info !== nextProps.info;
+  },
+
   render: function() {
-    var ship = this.props.ship;
-    var info = ship.info;
     return (
       <div className="ship-info">
         <h2>Ship Info</h2>
-        {this.renderElement('shipName', info.shipName, 'Ship Name')}
-        {this.renderElement('captain', info.captain, 'Captain Name')}
-        {this.renderElement('firstOfficer', info.firstOfficer, 'First Officer')}
-        {this.renderElement('chiefEngineer', info.chiefEngineer, 'Chief Engineer')}
-        {this.renderElement('tacticalOfficer', info.tacticalOfficer, 'Tactical Officer')}
-        {this.renderElement('helmsman', info.helmsman, 'Helmsman')}
+        {this.renderElements(this.props.info)}
       </div>
     );
   },
 
-  renderElement: function(keyName, item, defaultValue) {
-    return (
-      <EditableElement
-          keyName={keyName}
-          item={item}
-          defaultValue={defaultValue}
-          onEdit={this.updateShipInfo}/>
-    );
+  renderElements: function(info) {
+    return Object.keys(info).map(function(key, index) {
+      return (
+        <EditableElement
+          value={this.getValue(key)}
+          key={index}
+          onEdit={this.updateInfo.bind(this, key)} />
+      );
+    }.bind(this));
   },
 
-  updateShipInfo: function(key, newValue) {
-    var ship = this.props.ship;
-    ship.info[key] = newValue;
-    this.props.updateShip(ship);
+  updateInfo: function(key, newValue) {
+    var info = this.props.info;
+    info[key] = newValue;
+    this.props.updateShipInfo(info);
+  },
+
+  getValue: function(key) {
+    return this.props.info[key] || this.keyToValue(key);
+  },
+
+  keyToValue: function(key) {
+    // insert a space before all caps and upper case the first character
+    return key.replace(/([A-Z])/g, ' $1')
+      .replace(/^./, function(str){ return str.toUpperCase(); })
   }
 });
-
-module.exports = ShipInfo;
