+++ finished/app/components/WarpDriveControls.jsx
@@ -1,18 +1,17 @@
-var React             = require('react');
-var nav               = require('../utilities/starshipNavigation.js');
-var EditableElement   = require('./EditableElement.jsx');
+var React = require('react');
+var nav = require('../utilities/starshipNavigation.js');
+var EditableElement = require('./EditableElement.jsx');
 
-var WarpDriveControls = React.createClass({
+module.exports = React.createClass({
+
+  displayName: "WarpDriveControls",
+
   render: function() {
     return (
       <div className="navigational-controls">
         <h2>Warp Drive</h2>
         <span>WARP:</span>
-        <EditableElement
-          keyName="name"
-          item={this.props.ship.speed}
-          defaultValue="Full Stop"
-          onEdit={this.updateSpeed}/>
+        <EditableElement value={this.props.speed} onEdit={this.updateSpeed}/>
         <div className="arrow-controls">
           <button onClick={this.incrementSpeed.bind(this, 1)}>&#11014;</button>
           <button onClick={this.incrementSpeed.bind(this, -1)}>&#11015;</button>
@@ -24,18 +23,18 @@ var WarpDriveControls = React.createClass({
     );
   },
 
-  updateSpeed: function(keyName, newValue) {
-    var newSpeed = parseInt(newValue, 10);
-    this.props.updateSpeed(newSpeed);
+  updateSpeed: function(newSpeed) {
+    var newSpeed = parseFloat(newSpeed, 10);
+    if (0 <= newSpeed && newSpeed < 10) {
+      this.props.updateSpeed(newSpeed);
+    }
   },
 
   incrementSpeed: function(dSpeed) {
-    var currentSpeed = this.props.ship.speed;
+    var currentSpeed = this.props.speed;
     var newSpeed = Math.floor(currentSpeed + dSpeed);
     if (0 <= newSpeed && newSpeed < 10) {
       this.props.updateSpeed(newSpeed);
     }
   }
 });
-
-module.exports = WarpDriveControls;
