+++ finished/app/data/StarData.js
@@ -1,0 +1,1022 @@
+module.exports = [
+  // Row 1 of Map
+  {
+    name: 'Breen',
+    position: [10, 90],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Kul Tiras',
+    position: [160, 95],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Celes',
+    position: [170, 70],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Alterac',
+    position: [220, 40],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'El Toris',
+    position: [260, 90],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Azel',
+    position: [310, 20],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Arbazan',
+    position: [340, 10],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Starbase 204',
+    position: [345, 80],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Theta Minsar',
+    position: [350, 90],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Theta Cygni',
+    position: [350, 60],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Byzell',
+    position: [390, 20],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Corinth',
+    position: [435, 35],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Delta',
+    position: [440, 90],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Kilka',
+    position: [505, 45],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Lyrata',
+    position: [530, 50],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Ka\'Tula',
+    position: [550, 70],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Starbase 241',
+    position: [570, 90],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Gildebron',
+    position: [590, 40],
+    jurisdiction: 'alteran-expanse'
+  },
+  {
+    name: 'Ciden',
+    position: [630, 90],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Cheron',
+    position: [650, 70],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Romulus',
+    position: [670, 90],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Rho Tucanae',
+    position: [840, 70],
+    jurisdiction: 'garidian-republic'
+  },
+  {
+    name: 'Cortaynus',
+    position: [860, 40],
+    jurisdiction: 'garidian-republic'
+  },
+  {
+    name: 'Gacrux',
+    position: [870, 80],
+    jurisdiction: 'garidian-republic'
+  },
+  {
+    name: 'T\'Met',
+    position: [895, 30],
+    jurisdiction: 'garidian-republic'
+  },
+  {
+    name: 'T\'Bak',
+    position: [950, 70],
+    jurisdiction: 'garidian-republic'
+  },
+
+  // Row 2 of Map
+
+  {
+    name: 'Nathrazim',
+    position: [40, 110],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Adin Kel',
+    position: [140, 140],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Tirsfal',
+    position: [170, 180],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Kaldir',
+    position: [170, 110],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Nahal Kaitos',
+    position: [205, 120],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'J\'Larre',
+    position: [240, 150],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Tethe-Alla',
+    position: [240, 190],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Amizor',
+    position: [290, 170],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Starbase 54',
+    position: [330, 140],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Corneria',
+    position: [350, 105],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Terminia',
+    position: [390, 130],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Ohniaka',
+    position: [460, 105],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Zetar',
+    position: [440, 170],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Sauria',
+    position: [480, 120],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Ivor',
+    position: [510, 150],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Lukaris',
+    position: [540, 140],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Bolarus',
+    position: [550, 170],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Aurillac',
+    position: [595, 110],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Dessica',
+    position: [595, 180],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Devron',
+    position: [605, 195],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Trikkala',
+    position: [710, 190],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Celinar',
+    position: [760, 150],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Haakona',
+    position: [840, 130],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Dimovius',
+    position: [905, 170],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Pretorian',
+    position: [930, 190],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Laraq',
+    position: [920, 110],
+    jurisdiction: 'romulan-empire'
+  },
+
+  // Row 3 of Map
+  {
+    name: 'Reza-Karun',
+    position: [40, 210],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Matherital',
+    position: [90, 280],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Shengra Nel',
+    position: [170, 240],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Tukatu',
+    position: [190, 260],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Maxia',
+    position: [280, 220],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Forren',
+    position: [290, 260],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Tarkalia',
+    position: [330, 210],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Denobula',
+    position: [390, 280],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Vega',
+    position: [390, 250],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Inferna',
+    position: [390, 230],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Tellar',
+    position: [440, 290],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Memory Alpha',
+    position: [460, 210],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Sol',
+    position: [500, 300],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Alpha Centauri',
+    position: [520, 280],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Draylax',
+    position: [530, 250],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Babel',
+    position: [550, 290],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Benzar',
+    position: [550, 220],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Kolaran',
+    position: [610, 230],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Calder',
+    position: [630, 240],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Oscelliet',
+    position: [640, 250],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Draken',
+    position: [640, 270],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Kaleb',
+    position: [650, 280],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Nelvana',
+    position: [680, 270],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Alinor',
+    position: [690, 280],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Talon',
+    position: [730, 230],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Vendor',
+    position: [740, 250],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Glintara',
+    position: [780, 290],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Algorab',
+    position: [790, 250],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Achenar',
+    position: [840, 240],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Penthara',
+    position: [890, 240],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Cheltax',
+    position: [930, 270],
+    jurisdiction: 'romulan-empire'
+  },
+
+  // Row 4 of Map
+
+  {
+    name: 'Sylvarrant',
+    position: [70, 320],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Voltar',
+    position: [40, 340],
+    jurisdiction: 'elorg-bloc'
+  },
+  {
+    name: 'Felton',
+    position: [20, 390],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Athos',
+    position: [130, 400],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Delof',
+    position: [130, 350],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Alpha 441',
+    position: [170, 390],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Lamenda',
+    position: [190, 350],
+    jurisdiction: 'tzenkethi-coalition'
+  },
+  {
+    name: 'Tzenketh',
+    position: [270, 350],
+    jurisdiction: 'tzenkethi-coalition'
+  },
+  {
+    name: 'Temecklia',
+    position: [290, 390],
+    jurisdiction: 'tzenkethi-coalition'
+  },
+  {
+    name: 'Barisa',
+    position: [300, 370],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'AR-558',
+    position: [330, 330],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Mizar',
+    position: [340, 380],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Sira Adun',
+    position: [370, 390],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Terra Nova',
+    position: [430, 350],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Starbase 74',
+    position: [460, 390],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Axanar',
+    position: [480, 320],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Wolf 359',
+    position: [530, 320],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Andoria',
+    position: [535, 340],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Vulcan',
+    position: [530, 360],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Weytahn',
+    position: [600, 350],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Efrosia',
+    position: [610, 400],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Horst',
+    position: [640, 370],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Taugus',
+    position: [670, 330],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Qualor',
+    position: [710, 340],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Jolanisar',
+    position: [720, 310],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Jantral',
+    position: [720, 370],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Algeron',
+    position: [750, 330],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Galorndon Core',
+    position: [750, 340],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Tomed',
+    position: [770, 370],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Outpost 23',
+    position: [820, 390],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Alpha Onias',
+    position: [840, 360],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Uniroth',
+    position: [860, 370],
+    jurisdiction: 'romulan-empire'
+  },
+  {
+    name: 'Al\'Asayan',
+    position: [940, 380],
+    jurisdiction: 'romulan-empire'
+  },
+
+  // Row 5 of Map
+
+  {
+    name: 'Lazon',
+    position: [10, 440],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Actrium',
+    position: [80, 410],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Adarak',
+    position: [60, 450],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Kora',
+    position: [70, 480],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Quinor',
+    position: [90, 430],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Monac',
+    position: [110, 480],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Torros',
+    position: [130, 490],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Marva',
+    position: [170, 420],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Olmerak',
+    position: [180, 430],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Cardassia',
+    position: [190, 460],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Empok Nor',
+    position: [220, 410],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Korma',
+    position: [210, 480],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Deep Space 9',
+    position: [240, 440],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Bajor',
+    position: [250, 460],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Kobliad',
+    position: [290, 410],
+    jurisdiction: 'tzenkethi-coalition'
+  },
+  {
+    name: 'Loval',
+    position: [220, 490],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Xepolite',
+    position: [280, 480],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Lyshan',
+    position: [305, 460],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Trill',
+    position: [330, 420],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Kabrel',
+    position: [340, 440],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Kalandra',
+    position: [360, 460],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Quatal',
+    position: [350, 490],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Lya Station Alpha',
+    position: [380, 480],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Kressari',
+    position: [390, 470],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Draylon',
+    position: [410, 430],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Bynaus',
+    position: [440, 450],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Orion',
+    position: [530, 420],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Coridan',
+    position: [520, 440],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Rigel',
+    position: [590, 450],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Nausicaa',
+    position: [640, 460],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Regulus',
+    position: [680, 480],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Celes',
+    position: [690, 420],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Midrian',
+    position: [720, 490],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Japori',
+    position: [790, 460],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Veytan',
+    position: [810, 440],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Jouret',
+    position: [820, 420],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Starbase 234',
+    position: [840, 480],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Acmar',
+    position: [860, 440],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Stabase 84',
+    position: [890, 430],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Nimbus',
+    position: [940, 440],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Chi\'Dan',
+    position: [960, 480],
+    jurisdiction: 'klingon-empire'
+  },
+
+  // Row 6 of Map
+
+  {
+    name: 'Dopa',
+    position: [20, 540],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Sarpedion',
+    position: [70, 510],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Orias',
+    position: [90, 520],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Velos',
+    position: [110, 570],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Pentath',
+    position: [150, 510],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Soukara',
+    position: [170, 530],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Juhraya',
+    position: [160, 570],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Tracken',
+    position: [180, 590],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Omekla',
+    position: [190, 550],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Celtris',
+    position: [190, 540],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Amleth',
+    position: [220, 520],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Rakal',
+    position: [240, 510],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Aschelon',
+    position: [260, 540],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Kelrabi',
+    position: [270, 570],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Argus Array',
+    position: [290, 590],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Chin\'Toka',
+    position: [340, 520],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Setlik',
+    position: [320, 570],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Cuellar',
+    position: [330, 550],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Veloz',
+    position: [360, 540],
+    jurisdiction: 'cardassian-union'
+  },
+  {
+    name: 'Solarion',
+    position: [390, 520],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Argolis',
+    position: [390, 510],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Minos Korva',
+    position: [380, 550],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Malcor',
+    position: [420, 580],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Betazed',
+    position: [440, 540],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Capella',
+    position: [445, 500],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Dakala',
+    position: [450, 520],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Tandar',
+    position: [480, 560],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Mazar',
+    position: [520, 505],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Risa',
+    position: [540, 520],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Castor',
+    position: [545, 540],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Elora',
+    position: [560, 580],
+    jurisdiction: 'federation'
+  },
+  {
+    name: 'Son\'A',
+    position: [590, 590],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Tarlac',
+    position: [590, 580],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Baku',
+    position: [590, 550],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Kendra',
+    position: [615, 520],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Suliban',
+    position: [640, 540],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Obruli',
+    position: [670, 590],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Turkana',
+    position: [710, 570],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Yaraka',
+    position: [740, 540],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Verix',
+    position: [770, 580],
+    jurisdiction: 'independent'
+  },
+  {
+    name: 'Khitomer',
+    position: [880, 540],
+    jurisdiction: 'klingon-empire'
+  },
+  {
+    name: 'Morska',
+    position: [890, 580],
+    jurisdiction: 'klingon-empire'
+  },
+  {
+    name: 'Rura Penthe',
+    position: [910, 590],
+    jurisdiction: 'klingon-empire'
+  },
+  {
+    name: 'Qo\' Nos',
+    position: [960, 580],
+    jurisdiction: 'klingon-empire'
+  }
+];
