+++ finished/app/main.css
@@ -71,6 +71,21 @@ svg {
   border-radius: 10px;
   font-family: monospace;
   font-size: 16px;
+  padding-bottom: 20px;
+}
+
+.helm > div:not(#helm-header) {
+  display: inline-block;
+  width: 200px;
+  border: 2px solid #B3C8FD;
+  margin-top: 0;
+  padding-top: 0;
+  border-radius: 10px;
+  vertical-align: top;
+  margin: 0 10px;
+  background-color: #2D4E94;
+  color: #ECE660;
+  height: 315px;
 }
 
 .helm h2 {
@@ -84,37 +99,6 @@ svg {
   color: #182C5C;
   padding-right: 25px;
   text-shadow: none;
-}
-
-.ship-info {
-  display: inline-block;
-  width: 200px;
-  border: 2px solid #B3C8FD;
-  border-radius: 10px;
-  vertical-align: top;
-  margin: 0 10px 20px 10px;
-  background-color: #2D4E94;
-  color: #ECE660;
-  height: 315px;
-}
-
-.navigation {
-  display: inline-block;
-  vertical-align: top;
-}
-
-.navigation > div {
-  display: inline-block;
-  width: 200px;
-  border: 2px solid #B3C8FD;
-  margin-top: 0;
-  padding-top: 0;
-  border-radius: 10px;
-  vertical-align: top;
-  margin: 0 10px;
-  background-color: #2D4E94;
-  color: #ECE660;
-  height: 315px;
 }
 
 .arrow-controls {
@@ -171,8 +155,8 @@ svg {
 
 .ui-state-active,
 .ui-widget-content .ui-state-active,
-.ui-widget-header .ui-state-active, 
-.ui-autocomplete, .ui-autocomplete:hover, 
+.ui-widget-header .ui-state-active,
+.ui-autocomplete, .ui-autocomplete:hover,
 .ui-menu-item, .ui-menu-item:hover,
 .ui-menu-item a, .ui-menu-item a:hover,
 .ui-widget-content .ui-state-focus,
