+++ finished/app/mixins/SetIntervalMixin.jsx
@@ -1,19 +1,18 @@
-var SetIntervalMixin = {
+module.exports = {
   componentWillMount: function() {
     this.intervals = [];
-  },
-
-  setInterval: function() {
-    this.intervals.push(setInterval.apply(null, arguments));
   },
 
   componentWillUnmount: function() {
     this.clearIntervals();
   },
 
+  setInterval: function() {
+    this.intervals.push(setInterval.apply(null, arguments));
+  },
+
   clearIntervals: function() {
     this.intervals.forEach(clearInterval);
+    this.intervals = [];
   }
 };
-
-module.exports = SetIntervalMixin;
