--- unfinished/app/components/App.jsx
@@ -1,7 +1,0 @@
-var React             = require('react');
-var Stars             = require('../data/Stars.js');
-var StarChart         = require('./StarChart.jsx');
-var ShipInfo          = require('./ShipInfo.jsx');
-var Navigation        = require('./Navigation.jsx');
-var SetIntervalMixin  = require('../mixins/SetIntervalMixin.jsx');
-var nav               = require('../utilities/starshipNavigation.js');
