--- unfinished/app/components/CourseControl.jsx
@@ -1,4 +1,0 @@
-require('jquery-ui/autocomplete');
-var React    = require('react');
-var ReactDOM = require('react-dom');
-var utils    = require('../utilities/starUtilities.js');
