--- unfinished/app/components/EditableElement.jsx
@@ -1,1 +1,0 @@
-var React = require('react');
