--- unfinished/app/components/Navigation.jsx
@@ -1,4 +1,0 @@
-var React                = require('react');
-var NavigationDashboard  = require('./NavigationDashboard.jsx');
-var WarpDriveControls    = require('./WarpDriveControls.jsx');
-var CourseControl        = require('./CourseControl.jsx');
