--- unfinished/app/components/NavigationDashboard.jsx
@@ -1,2 +1,0 @@
-var React = require('react');
-var nav   = require('../utilities/starshipNavigation.js');
