--- unfinished/app/components/ShipInfo.jsx
@@ -1,2 +1,0 @@
-var React            = require('react');
-var EditableElement  = require('./EditableElement.jsx');
