--- unfinished/app/components/StarChart.jsx
@@ -1,2 +1,0 @@
-var React    = require('react');
-var Starship = require('./Starship.jsx');
