--- unfinished/app/components/Starship.jsx
@@ -1,2 +1,0 @@
-var React         = require('react');
-var starshipImage = require('../utilities/starshipImage.js');
