--- unfinished/app/components/WarpDriveControls.jsx
@@ -1,3 +1,0 @@
-var React             = require('react');
-var nav               = require('../utilities/starshipNavigation.js');
-var EditableElement   = require('./EditableElement.jsx');
