--- unfinished/app/index.jsx
@@ -1,4 +1,0 @@
-require('./main.css');
-var React    = require('react');
-var ReactDOM = require('react-dom');
-var App      = require('./components/App.jsx');
