--- unfinished/app/mixins/SetIntervalMixin.jsx
@@ -1,19 +1,0 @@
-var SetIntervalMixin = {
-  componentWillMount: function() {
-    this.intervals = [];
-  },
-
-  setInterval: function() {
-    this.intervals.push(setInterval.apply(null, arguments));
-  },
-
-  componentWillUnmount: function() {
-    this.clearIntervals();
-  },
-
-  clearIntervals: function() {
-    this.intervals.forEach(clearInterval);
-  }
-};
-
-module.exports = SetIntervalMixin;
