+++ unfinished/package.json
@@ -1,12 +1,11 @@
 {
   "name": "react-under-the-hood",
   "version": "1.0.0",
-  "description": "This is part of the React Under the Hood workshop",
-  "main": "index.js",
+  "description": "This is part of the React Under the Hood book",
+  "main": "index.jsx",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server",
-    "lint": "eslint . --ext .js --ext .jsx"
+    "start": "webpack-dev-server"
   },
   "repository": {
     "type": "git",
@@ -26,17 +25,12 @@
   "homepage": "https://github.com/freddyrangel/react-under-the-hood#readme",
   "devDependencies": {
     "babel-core": "^5.8.25",
-    "babel-eslint": "^4.1.3",
     "babel-loader": "^5.3.2",
     "css-loader": "^0.18.0",
-    "eslint": "^1.5.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-react": "^3.4.2",
     "file-loader": "^0.8.4",
     "html-webpack-plugin": "^1.6.1",
     "node-libs-browser": "^0.5.3",
-    "node-uuid": "^1.4.3",
-    "react-hot-loader": "^1.3.0",
+    "react-addons-perf": "^0.14.3",
     "style-loader": "^0.12.4",
     "webpack": "^1.12.2",
     "webpack-dev-server": "^1.11.0",
