+++ unfinished/webpack.config.js
@@ -6,7 +6,7 @@ var merge             = require('webpack-merge');
 var TARGET    = process.env.npm_lifecycle_event;
 var ROOT_PATH = path.resolve(__dirname);
 
-var common = {
+module.exports = {
   entry: path.resolve(ROOT_PATH, 'app'),
   resolve: {
     extensions: ['', '.js', '.jsx']
@@ -15,6 +15,7 @@ var common = {
     path: path.resolve(ROOT_PATH, 'build'),
     filename: 'bundle.js'
   },
+  devtool: 'eval-source-map',
   module: {
     loaders: [
       {
@@ -22,7 +23,16 @@ var common = {
         loaders: ['style', 'css'],
         include: path.resolve(ROOT_PATH, 'app')
       },
-      { test: /\.(ttf|eot|svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/, loader: "file-loader" }
+      {
+        test: /\.jsx?$/,
+        loaders: ['babel'],
+        include: path.resolve(ROOT_PATH, 'app')
+      },
+      {
+        test: /\.(ttf|eot|svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
+        loaders: ["file-loader"],
+        include: path.resolve(ROOT_PATH, 'app')
+      }
     ]
   },
   plugins: [
@@ -33,31 +43,11 @@ var common = {
       $: "jquery",
       jQuery: "jquery"
     })
-  ]
+  ],
+  devServer: {
+    historyApiFallback: true,
+    inline: true,
+    progress: true,
+    port: 4000
+  }
 };
-
-if (TARGET === 'start' || !TARGET) {
-  module.exports = merge(common, {
-    devtool: 'eval-source-map',
-    module: {
-      loaders: [
-        {
-          test: /\.jsx?$/,
-          loaders: ['react-hot', 'babel', 'eslint'],
-          include: path.resolve(ROOT_PATH, 'app')
-        }
-      ]
-    },
-    devServer: {
-      historyApiFallback: true,
-      hot: true,
-      inline: true,
-      progress: true,
-      port: 4000
-    },
-    plugins: [
-      new webpack.HotModuleReplacementPlugin()
-    ]
-  })
-}
-
